Modified RoleId data type in Model

+++ CT.Web/Common/CustomPrincipalSerializeModel.cs
@@ -7,7 +7,7 @@ namespace CT.Web.Common
 {
     public class CustomPrincipalSerializeModel
     {
-        public Int64 RoleId { get; set; }
+        public int RoleId { get; set; }
         public Int64 UserId { get; set; }
         public string UserName { get; set; }
         public string ProfileName { get; set; }
